+++ prime-router/src/main/kotlin/transport/RESTTransport.kt
@@ -12,7 +12,7 @@ import gov.cdc.prime.router.TransportType
 import gov.cdc.prime.router.azure.ActionHistory
 import gov.cdc.prime.router.azure.WorkflowEngine
 import gov.cdc.prime.router.azure.db.enums.TaskAction
-import gov.cdc.prime.router.azure.observability.event.ReportEventService
+import gov.cdc.prime.router.azure.observability.event.IReportEventService
 import gov.cdc.prime.router.credentials.CredentialHelper
 import gov.cdc.prime.router.credentials.CredentialRequestReason
 import gov.cdc.prime.router.credentials.RestCredential
@@ -90,7 +90,7 @@ class RESTTransport(private val httpClient: HttpClient? = null) : ITransport {
         retryItems: RetryItems?,
         context: ExecutionContext,
         actionHistory: ActionHistory,
-        reportEventService: ReportEventService,
+        reportEventService: IReportEventService,
     ): RetryItems? {
         val logger: Logger = context.logger
